requirements.txt
pytelegrambotapi==4.14.0
requests==2.31.0
beautifulsoup4==4.12.2
google-generativeai==0.3.2